backend/requirements.txt
fastapi
uvicorn
httpx
requests
python-dotenv
google-auth
pydantic[email]
motor
pymongo
